+++ R/tweak-page.R
@@ -29,6 +29,7 @@ tweak_rmarkdown_html <- function(html, input_path, pkg = pkg) {
     )
   }
 
+  tweak_img_src(html)
   tweak_tables(html)
 
   invisible()
@@ -79,13 +80,8 @@ tweak_homepage_html <- function(html,
         .where = "before"
       )
   }
-  # Fix relative image links
-  imgs <- xml2::xml_find_all(html, ".//img")
-  urls <- xml2::xml_attr(imgs, "src")
-  new_urls <- gsub("^vignettes/", "articles/", urls)
-  new_urls <- gsub("^man/figures/", "reference/figures/", new_urls)
-  purrr::map2(imgs, new_urls, ~ (xml2::xml_attr(.x, "src") <- .y))
 
+  tweak_img_src(html)
   tweak_tables(html)
 
   invisible()