+++ tests/testthat/test-build-articles.R
@@ -5,6 +5,26 @@ test_that("can recognise intro variants", {
   expect_true(article_is_intro("articles/pack-age", "pack.age"))
 })
 
+test_that("links to man/figures are automatically relocated", {
+  pkg <- test_path("assets/man-figures")
+  dst <- withr::local_tempdir()
+
+  expect_output(build_articles(pkg, override = list(destination = dst)))
+
+  html <- xml2::read_html(path(dst, "articles", "kitten.html"))
+  src <- xpath_attr(html, "//img", "src")
+
+  expect_equal(src, c(
+    "../reference/figures/kitten.jpg",
+    "../reference/figures/kitten.jpg",
+    "another-kitten.jpg",
+    "https://www.tidyverse.org/rstudio-logo.svg"
+  ))
+
+  # And files aren't copied
+  expect_false(dir_exists(path(dst, "man")))
+})
+
 test_that("articles don't include header-attrs.js script", {
   pkg <- as_pkgdown(test_path("assets/articles"))
   withr::defer(clean_site(pkg))